pyproject.toml
[project]
name = "mcp-sandbox-env"
version = "0.1.0"
description = "MCP-Based Sandboxed Task Environment for AI Agent RL"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.0.0",
    "odfpy>=1.4.1",
    "openpyxl>=3.1.0",
    "pydantic>=2.0.0",
    "uvicorn>=0.27.0",
    "starlette>=0.36.0",
    "mcp>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "httpx>=0.27.0",
]

[tool.setuptools.packages.find]
include = ["mcp_server*"]
